requirements.txt
Pillow==10.0.0
requests==2.31.0
pycountry==22.3.5